- Changed folder structure; - Edited screen files to match size.
