+++ lib/screens/desktop/loginscreen_desktop.dart
@@ -1,0 +1,156 @@
+import 'package:boksklapps/all_imports.dart';
+
+/// LoginScreenDesktop class
+class LoginScreenDesktop extends StatefulWidget {
+  /// LoginScreenDesktop constructor
+  const LoginScreenDesktop({
+    super.key,
+  });
+
+  @override
+  State<LoginScreenDesktop> createState() => LoginScreenDesktopState();
+}
+
+/// LoginScreenDesktopState class
+class LoginScreenDesktopState extends State<LoginScreenDesktop> {
+  /// EmailController
+  final TextEditingController emailCtrl = TextEditingController();
+
+  /// PasswordController
+  final TextEditingController passwordCtrl = TextEditingController();
+
+  @override
+  void dispose() {
+    emailCtrl.dispose();
+    passwordCtrl.dispose();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Padding(
+        padding: EdgeInsets.fromLTRB(
+          MediaQuery.of(context).size.width / 4,
+          64,
+          MediaQuery.of(context).size.width / 4,
+          56,
+        ),
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.spaceBetween,
+          children: [
+            Column(
+              children: [
+                Row(
+                  mainAxisAlignment: MainAxisAlignment.end,
+                  children: <InkWell>[
+                    InkWell(
+                      onTap: () {
+                        Navigator.pushNamed(
+                          context,
+                          '/register_screen',
+                        );
+                      },
+                      child: const Text(
+                        StringUtils.kCreateAccount,
+                        style: TextStyleUtils.kHeadline3,
+                      ),
+                    ),
+                  ],
+                ),
+                const SizedBox(height: 24),
+                Row(
+                  mainAxisAlignment: MainAxisAlignment.end,
+                  children: <InkWell>[
+                    InkWell(
+                      onTap: () {
+                        Navigator.pushNamed(
+                          context,
+                          '/password_screen',
+                        );
+                      },
+                      child: const Text(
+                        StringUtils.kForgotPassword,
+                        style: TextStyleUtils.kHeadline3,
+                      ),
+                    ),
+                  ],
+                ),
+              ],
+            ),
+            const Spacer(),
+            Padding(
+              padding: const EdgeInsets.only(bottom: 48),
+              child: Column(
+                children: [
+                  const Text(
+                    'LOGIN',
+                    style: TextStyleUtils.kHeadline1,
+                  ),
+                  const SizedBox(height: 24),
+                  TextField(
+                    controller: emailCtrl,
+                    keyboardType: TextInputType.emailAddress,
+                    textAlign: TextAlign.center,
+                    decoration: const InputDecoration(
+                      labelText: StringUtils.kLabelEmail,
+                      prefixIcon: Icon(
+                        Icons.email_outlined,
+                      ),
+                    ),
+                  ),
+                  const SizedBox(height: 12),
+                  TextField(
+                    controller: passwordCtrl,
+                    keyboardType: TextInputType.text,
+                    obscureText: true,
+                    textAlign: TextAlign.center,
+                    enableSuggestions: false,
+                    decoration: const InputDecoration(
+                      labelText: StringUtils.kLabelPassword,
+                      prefixIcon: Icon(
+                        Icons.lock_outlined,
+                      ),
+                    ),
+                  ),
+                ],
+              ),
+            ),
+          ],
+        ),
+      ),
+      floatingActionButton: FloatingActionButton(
+        onPressed: () async {
+          if (emailCtrl.text.isEmpty) {
+            ScaffoldMessenger.of(context).showSnackBar(
+              const SnackBar(
+                content: Text('Email cannot be empty'),
+              ),
+            );
+            return;
+          } else if (passwordCtrl.text.isEmpty) {
+            ScaffoldMessenger.of(context).showSnackBar(
+              const SnackBar(
+                content: Text('Password cannot be empty'),
+              ),
+            );
+            return;
+          } else if (passwordCtrl.text.length < 8) {
+            ScaffoldMessenger.of(context).showSnackBar(
+              const SnackBar(
+                content: Text('Password must be at least 8 characters'),
+              ),
+            );
+            return;
+          } else {
+            await Navigator.pushReplacementNamed(
+              context,
+              '/home_screen',
+            );
+          }
+        },
+        child: IconUtils.kForward,
+      ),
+    );
+  }
+}
