+++ lib/screens/home_screen.dart
@@ -14,73 +14,3 @@ class HomeScreen extends StatelessWidget {
     }
   }
 }
-
-/// HomeScreenMobile class
-class HomeScreenMobile extends StatelessWidget {
-  /// HomeScreenMobile constructor
-  const HomeScreenMobile({
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text(
-          StringUtils.kBoks + StringUtils.kKlapps,
-        ),
-        centerTitle: true,
-      ),
-      // Drawer hides if the screen width is less than 1000
-      drawer: const DrawerWidget(),
-      body: Column(
-        children: const [
-          GetShitDoneWidget(),
-          TimerWidget(),
-        ],
-      ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: () {},
-        child: IconUtils.kForward,
-      ),
-    );
-  }
-}
-
-/// HomeScreenDesktop class
-class HomeScreenDesktop extends StatelessWidget {
-  /// HomeScreenDesktop constructor
-  const HomeScreenDesktop({
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text(
-          StringUtils.kBoks + StringUtils.kKlapps,
-        ),
-        centerTitle: true,
-      ),
-      body: Row(
-        children: [
-          // Drawer shows if the screen width is greater than 1000
-          const DrawerWidget(),
-          Expanded(
-            child: Column(
-              children: const [
-                GetShitDoneWidget(),
-                TimerWidget(),
-              ],
-            ),
-          ),
-        ],
-      ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: () {},
-        child: IconUtils.kForward,
-      ),
-    );
-  }
-}
