+++ lib/screens/settings_screen.dart
@@ -1,42 +1,17 @@
 import 'package:boksklapps/all_imports.dart';
 
 /// SettingsScreen class
-class SettingsScreen extends ConsumerWidget {
+class SettingsScreen extends StatelessWidget {
   /// SettingsScreen constructor
   const SettingsScreen({super.key});
 
   @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text(
-          StringUtils.kSettings,
-        ),
-        centerTitle: true,
-      ),
-      body: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          ListTile(
-            title: const Text(
-              StringUtils.kLightSystemDark,
-            ),
-            subtitle: Text(
-              ref.watch(themeModeStringProvider),
-            ),
-            trailing: const ThemeModeSwitch(),
-          ),
-          ListTile(
-            title: const Text(
-              StringUtils.kColorScheme,
-            ),
-            subtitle: Text(
-              ref.watch(themeColorStringProvider),
-            ),
-            trailing: const ThemeColorSwitch(),
-          ),
-        ],
-      ),
-    );
+  Widget build(BuildContext context) {
+    if (MediaQuery.of(context).size.width < 1000) {
+      return const SettingsScreenMobile();
+    } else {
+      //TODO: Add Desktop version of SettingsScreen()
+      return const SettingsScreenMobile();
+    }
   }
 }
